requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
discord.py==1.4.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.9
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.1
multidict==4.7.6
numpy==1.19.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.1
urllib3==1.25.10
yarl==1.5.1